requirements.txt
google-genai>=0.8.0
flask>=3.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
